requirements.txt
streamlit
htmlmin
numpy
pandas
matplotlib
jinja2
sqlalchemy
python-dotenv